Unit Root Tests for Seasonal Time Series

Canova and Hansen (CH) test for seasonal stability and Hylleberg, Engle, Granger and Yoo (HEGY) test for seasonal unit roots.

Details

Version >= 2.0.0 is a revival of the former package uroot. Some of the functions provided in the original versions have been coded from the scratch.

The current version provides the Canova and Hansen (CH) test for seasonal stability and the Hylleberg, Engle, Granger and Yoo (HEGY) test for seasonal unit roots.

New features:

The original functions have been enhanced with the following new features: 1) the tests are now applicable to series of any seasonal periodicity (not only quarterly and monthly data), 2) p-values based on response surface regressions are available, 3) bootstrapped p-values are available for the HEGY test statistics.

Old features not currently supported:

Some of the utilities available in the initial versions of the package are not available now: graphics for seasonal series, graphical user interface, bootstrap versions of the statistics, ADF test and KPSS test. Some of these utilities are available in other packages and some will probably be incorporated in future versions of this package.

System requirements:

Windows systems: GPU parallelization of the bootstrap is not currently operational on windows systems.

Unix systems: GPU parallelization of the bootstrap requires a CUDA capable GPU with compute capability >= 3.0.

References

Burridge, P. and Taylor, R. (2004) "Bootstrapping the HEGY seasonal unit root tests." Journal of Econometrics 123(1), pp. 67-87. DOI: http://dx.doi.org/10.1016/j.jeconom.2003.10.029.

Canova, F. and Hansen, Bruce E. (1995) "Are seasonal patterns constant over time? A test for seasonal stability". Journal of Business & Economic Statistics, 13(3), pp. 237-252. DOI: http://dx.doi.org/10.1080/07350015.1995.10524598.

Díaz-Emparanza, I. (2014) "Numerical distribution functions for seasonal unit root tests"- Computational Statistics and Data Analysis, 76, pp. 237-247. DOI: http://dx.doi.org/10.1016/j.csda.2013.03.006.

Díaz-Emparanza, I. and Moral, M. P. (2013). Seasonal stability tests in gretl. An application to international tourism data. Working paper: Biltoki D.T. 2013.03.

Hylleberg, S., Engle, R., Granger, C. and Yoo, B. (1990) "Seasonal integration and cointegration". Journal of Econometrics 44(1), pp. 215-238. DOI: http://dx.doi.org/10.1016/0304-4076(90)90080-D.