Original Tables of Critical Values

Compute p-values by interpolation in the tables of critical values provided in the original references given below.

uroot.raw.pvalue(x, type = c("CH", "HEGY"), v, n, ctd, S, Ftpi)

Arguments

x
a numeric. The value of the CH statistic.

type
a character specifying the type of test statistic.

v
numeric, the degrees of freedom of the Von Mises distribution. Only for type="CH".

n
numeric, the number of observations.

ctd
a character indicating the deterministic elements that were included in the HEGY regression. This argument is defined as paste(deterministic, collapse = ""), where deterministic is the argument of that name that was passed to hegy.test. Only for type="ADF" or type="HEGY".

S
numeric, the periodicity of the data.

Ftpi
a character indicating whether the type of statistic: "zero", \(t\)-test for the zero frequency; "pi", \(t\)-test for the frequency \(\pi\); "pair", \(F\)-test for the pairs of complex conjugates frequencies. Only for type="ADF" or type="HEGY".

Details

This function is used internally by ch.test and hegy.test.

Value

A numeric giving the calculated p-value.

See also

References

Beaulieu, J. J. Miron, J. A. (1993) "Seasonal Unit Roots in Aggregate U.S. Data". Journal of Econometrics, 55(1-2), pp. 305-328. DOI: http://dx.doi.org/10.1016/0304-4076(93)90018-Z.

Canova, F. and Hansen, Bruce E. (1995) "Are Seasonal Patterns Constant over Time? A Test for Seasonal Stability". Journal of Business & Economic Statistics, 13(3), pp. 237-252. DOI: http://dx.doi.org/10.1080/07350015.1995.10524598.

Hylleberg, S., Engle, R., Granger, C. and Yoo, B. (1990) "Seasonal integration and cointegration". Journal of Econometrics 44(1), pp. 215-238. DOI: http://dx.doi.org/10.1016/0304-4076(90)90080-D.